test for getLoomEmbedUrl
